+++ sdk/python/pulumi_aws/elasticloadbalancing/listener_policy.py
@@ -13,7 +13,7 @@ class ListenerPolicy(pulumi.CustomResource):
     """
     The load balancer to attach the policy to.
     """
-    load_balancer_port: pulumi.Output[int]
+    load_balancer_port: pulumi.Output[float]
     """
     The load balancer listener port to apply the policy to.
     """
@@ -28,7 +28,7 @@ class ListenerPolicy(pulumi.CustomResource):
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] load_balancer_name: The load balancer to attach the policy to.
-        :param pulumi.Input[int] load_balancer_port: The load balancer listener port to apply the policy to.
+        :param pulumi.Input[float] load_balancer_port: The load balancer listener port to apply the policy to.
         :param pulumi.Input[list] policy_names: List of Policy Names to apply to the backend server.
         """
         if __name__ is not None:
